Have the best of both worlds.
This provides a configuration options to enable and disable the ability to allow other branches. Such a feature is useful because the devop may want to enforce the strict deployment only nature of the repo OR they may use it to store other branches as it is the main git repository and anything in master gets deployed locally (e.g. to a staging server).